+++ Changelog.md
@@ -37,6 +37,8 @@ Bug Fixes:
   on an any instance. (Xavier Shay)
 * Fix `and_call_original` to work properly when multiple classes in an
   inheritance hierarchy have been stubbed with the same method. (Myron Marston)
+* Fix `any_instance` so that it updates existing instances that have
+  already been stubbed. (Myron Marston)
 
 ### 3.0.0.beta2 / 2014-02-17
 [Full Changelog](http://github.com/rspec/rspec-mocks/compare/v3.0.0.beta1...v3.0.0.beta2)